De-debianizing linux-flavour defaults on i386 and only including 486.
This is 'debranding' in the sense that live-build defaults are for people
building images with a config from scratch, live-build should not
have the options that we use for the official prebuilt debian-live
images as debian-defaults for all images and force that upon everyone
(as defaults). Rather, a 'lb config' should produce a as vanilla as
possible configuration tree.