Handling packages.chroot with an explicitly trusted file:// repository.
Up to now we created a temporary GPG key that we registered with apt-key
but with the switch to GnuPG 2 by default, this code broke. Now we stop
doing that but we add the “trusted=yes“ attribute in sources.list so
that APT knows that the repository can be trusted even if it's unsigned.

Signed-off-by: Daniel Baumann <daniel.baumann@open-infrastructure.net>