Fix including local packages.
Previously, including packages in config/chroot_local-packages would
break the Tails build.  The reason is that the version of live-build
used by Tails tried to create an OpenPGP key to sign the local
packages, but that key creation was designed with GnuPG 1.4.x in mind.
With 2.x, GnuPG tries to invoke a pinentry tool to ask for a
passworod, which failed likely due to stdin not being connected to a
terminal.

This patch drops the key generation, and instead marks the local apt
repository as trusted.  This is what contemporary versions of
live-build do.

Fixes https://gitlab.tails.boum.org/tails/tails/-/issues/15618